+++ Casks/moredock.rb
@@ -1,6 +1,6 @@
 cask "moredock" do
-  version "0.1.8"
-  sha256 "901cb15af2aac194a153b20e55048f196c88527b2db3df2a6650499bc3d91fe8"
+  version "0.1.9"
+  sha256 "aa5adb60b9ae70143405de425490c04d1ee6981227227e0bc86bd2cbb7bed650"
 
   url "https://github.com/ArioMoniri/moredock/releases/download/v#{version}/MoreDock-#{version}-macOS.dmg"
   name "MoreDock"